refactor: split admin diagnostic routes
